+++ lib/cask/source/path_base.rb
@@ -1,0 +1,25 @@
+class Cask::Source::PathBase
+
+  # derived classes must define method self.me?
+
+  def self.path_for_query(query)
+    path_string = "#{query}"
+    path_string.concat('.rb') unless path_string.match(%r{\.rb\Z}i)
+    Pathname.new(path_string)
+  end
+
+  attr_reader :path
+
+  def initialize(path)
+    @path = Pathname(path).expand_path
+  end
+
+  def load
+    require path
+    Cask.const_get(cask_class_name).new
+  end
+
+  def cask_class_name
+    path.basename.to_s.sub(/\.rb/, '').split('-').map(&:capitalize).join
+  end
+end
