+++ lib/cask/source/tapped.rb
@@ -25,6 +25,6 @@ class Cask::Source::Tapped
 
   def load
     path = self.class.path_for_query(title)
-    Cask::Source::Path.new(path).load
+    Cask::Source::PathSlashOptional.new(path).load
   end
 end
